Implement use_scratch for RDS storrs

+++ R/hash.R
@@ -73,9 +73,10 @@ serialize_to_raw <- function(x, ascii, xdr) {
 ## an object which cannot be retrieved later on, causing havoc
 ## upstream.
 write_serialized_rds <- function(value, filename, compress,
-                                 scratch_dir = NULL, long = 2^31 - 2) {
+                                 scratch_dir = NULL, long = 2^31 - 2,
+                                 use_scratch = TRUE) {
   withCallingHandlers(
-    try_write_serialized_rds(value, filename, compress, scratch_dir, long),
+    try_write_serialized_rds(value, filename, compress, scratch_dir, long, use_scratch),
     error = function(e) unlink(filename))
 }
 
@@ -84,9 +85,10 @@ write_serialized_rds <- function(value, filename, compress,
 ## close the connection on exit from try_write_serialized_rds and
 ## delete the file *after* that.
 try_write_serialized_rds <- function(value, filename, compress,
-                                     scratch_dir = NULL, long = 2^31 - 2) {
-  tmp <- tempfile(tmpdir = scratch_dir %||% tempdir())
+                                     scratch_dir = NULL, long = 2^31 - 2,
+                                     use_scratch = TRUE) {
 
+  tmp <- ifelse(use_scratch, tempfile(tmpdir = scratch_dir %||% tempdir()), filename)
   con <- (if (compress) gzfile else file)(tmp, "wb")
   needs_close <- TRUE
   on.exit(if (needs_close) close(con), add = TRUE)
@@ -99,16 +101,20 @@ try_write_serialized_rds <- function(value, filename, compress,
   }
   close(con)
   needs_close <- FALSE
-  file.rename(tmp, filename)
+  if (use_scratch) {
+    file.rename(tmp, filename)
+  }
 }
 
 
 ## Same pattern for write_lines.  The difference is that this will
 ## delete the key on a failed write (otherwise there's a copy
 ## involved)
-write_lines <- function(text, filename, ..., scratch_dir = NULL) {
+write_lines <- function(text, filename, ...,
+                        scratch_dir = NULL, use_scratch = TRUE) {
   withCallingHandlers(
-    try_write_lines(text, filename, ..., scratch_dir = scratch_dir),
+    try_write_lines(text, filename, ..., scratch_dir = scratch_dir,
+                    use_scratch = use_scratch),
     error = function(e) unlink(filename))
 }
 
@@ -116,10 +122,12 @@ write_lines <- function(text, filename, ..., scratch_dir = NULL) {
 ## This implements write-then-move for writeLines, which gives us
 ## atomic writes and rewrites.  If 'scratch' is on the same filesystem
 ## as dirname(filename), then the os's rename is atomic
-try_write_lines <- function(text, filename, ..., scratch_dir) {
-  tmp <- tempfile(tmpdir = scratch_dir %||% tempdir())
+try_write_lines <- function(text, filename, ..., scratch_dir, use_scratch) {
+  tmp <- ifelse(use_scratch, tempfile(tmpdir = scratch_dir %||% tempdir()), filename)
   writeLines(text, tmp, ...)
   ## Not 100% necessary and strictly makes this nonatomic
-  unlink(filename)
-  file.rename(tmp, filename)
+  if (use_scratch) {
+    unlink(filename)
+    file.rename(tmp, filename)
+  }
 }
